+++ renderer/renderer.go
@@ -5,6 +5,7 @@
 package renderer
 
 import (
+	"bytes"
 	"fmt"
 	urlPkg "net/url"
 	"regexp"
@@ -12,12 +13,24 @@ import (
 	"strings"
 
 	"code.rocketnine.space/tslocum/cview"
+	"github.com/alecthomas/chroma/formatters"
+	"github.com/alecthomas/chroma/lexers"
+	"github.com/alecthomas/chroma/styles"
 	"github.com/makeworld-the-better-one/amfora/config"
 	"github.com/spf13/viper"
 )
 
+// Terminal color information, set during display initialization by display/display.go
+var TermColor string
+
 // Regex for identifying ANSI color codes
 var ansiRegex = regexp.MustCompile(`\x1b\[[0-9;]*m`)
+
+// Regex for identifying possible language string, based on RFC 6838 and lexers used by Chroma
+var langRegex = regexp.MustCompile(`^([a-zA-Z0-9]+/)?[a-zA-Z0-9]+([a-zA-Z0-9!_\#\$\&\-\^\.\+]+)*`)
+
+// Regex for removing trailing newline (without disturbing ANSI codes) from code formatted with Chroma
+var trailingNewline = regexp.MustCompile(`(\r?\n)(?:\x1b\[[0-9;]*m)*$`)
 
 // RenderANSI renders plain text pages containing ANSI codes.
 // Practically, it is used for the text/x-ansi.
@@ -45,6 +58,10 @@ func RenderPlainText(s string) string {
 //
 // Set includeFirst to true if the prefix and suffix should be applied to the first wrapped line as well
 func wrapLine(line string, width int, prefix, suffix string, includeFirst bool) []string {
+	if width < 1 {
+		width = 1
+	}
+
 	// Anonymous function to allow recovery from potential WordWrap panic
 	var ret []string
 	func() {
@@ -183,7 +200,7 @@ func convertRegularGemini(s string, numLinks, width int, proxied bool) (string, 
 					// Add the link text in blue (in a region), and a gray link number to the left of it
 					// Those are the default colors, anyway
 
-					wrappedLink = wrapLine(linkText, width,
+					wrappedLink = wrapLine(linkText, width-indent,
 						strings.Repeat(" ", indent)+
 							`["`+strconv.Itoa(num-1)+`"][`+config.GetColorString("amfora_link")+`]`,
 						`[-][""]`,
@@ -198,8 +215,8 @@ func convertRegularGemini(s string, numLinks, width int, proxied bool) (string, 
 				} else {
 					// No color
 
-					wrappedLink = wrapLine(linkText, width,
-						strings.Repeat(" ", len(strconv.Itoa(num))+4)+ // +4 for spaces and brackets
+					wrappedLink = wrapLine(linkText, width-indent,
+						strings.Repeat(" ", indent)+ // +4 for spaces and brackets
 							`["`+strconv.Itoa(num-1)+`"]`,
 						`[""]`,
 						false, // Don't indent the first line, it's the one with link number
@@ -215,7 +232,7 @@ func convertRegularGemini(s string, numLinks, width int, proxied bool) (string, 
 				if viper.GetBool("a-general.color") {
 					// Color
 
-					wrappedLink = wrapLine(linkText, width,
+					wrappedLink = wrapLine(linkText, width-indent,
 						strings.Repeat(" ", indent)+
 							`["`+strconv.Itoa(num-1)+`"]`+linkTag,
 						`[-::-][""]`,
@@ -229,7 +246,7 @@ func convertRegularGemini(s string, numLinks, width int, proxied bool) (string, 
 				} else {
 					// No color
 
-					wrappedLink = wrapLine(linkText, width,
+					wrappedLink = wrapLine(linkText, width-indent,
 						strings.Repeat(" ", indent)+
 							`["`+strconv.Itoa(num-1)+`"]`,
 						`[::-][""]`,
@@ -248,7 +265,8 @@ func convertRegularGemini(s string, numLinks, width int, proxied bool) (string, 
 		} else if strings.HasPrefix(lines[i], "* ") {
 			if viper.GetBool("a-general.bullets") {
 				// Wrap list item, and indent wrapped lines past the bullet
-				wrappedItem := wrapLine(lines[i][1:], width,
+				wrappedItem := wrapLine(lines[i][1:],
+					width-4, // Subtract the 4 indent spaces
 					fmt.Sprintf("    [%s]", config.GetColorString("list_text")),
 					"[-]", false)
 				// Add bullet
@@ -256,7 +274,8 @@ func convertRegularGemini(s string, numLinks, width int, proxied bool) (string, 
 					wrappedItem[0] + "[-]"
 				wrappedLines = append(wrappedLines, wrappedItem...)
 			} else {
-				wrappedItem := wrapLine(lines[i][1:], width,
+				wrappedItem := wrapLine(lines[i][1:],
+					width-4, // Subtract the 4 indent spaces
 					fmt.Sprintf("    [%s]", config.GetColorString("list_text")),
 					"[-]", false)
 				// Add "*"
@@ -277,7 +296,9 @@ func convertRegularGemini(s string, numLinks, width int, proxied bool) (string, 
 				lines[i] = strings.TrimPrefix(lines[i], ">")
 				lines[i] = strings.TrimPrefix(lines[i], " ")
 				wrappedLines = append(wrappedLines,
-					wrapLine(lines[i], width, fmt.Sprintf("[%s::i]> ", config.GetColorString("quote_text")),
+					wrapLine(lines[i],
+						width-2, // Subtract 2 for width of prefix string
+						fmt.Sprintf("[%s::i]> ", config.GetColorString("quote_text")),
 						"[-::-]", true)...,
 				)
 			}
@@ -315,11 +336,46 @@ func RenderGemini(s string, width int, proxied bool) (string, []string) {
 	pre := false
 	buf := "" // Block of regular or preformatted lines
 
+	// Language, formatter, and style for syntax highlighting
+	lang := ""
+	formatterName := TermColor
+	styleName := viper.GetString("a-general.highlight_style")
+
 	// processPre is for rendering preformatted blocks
 	processPre := func() {
 
+		syntaxHighlighted := false
+
+		// Perform syntax highlighting if language is set
+		if lang != "" {
+			style := styles.Get(styleName)
+			if style == nil {
+				style = styles.Fallback
+			}
+			formatter := formatters.Get(formatterName)
+			if formatter == nil {
+				formatter = formatters.Fallback
+			}
+			lexer := lexers.Get(lang)
+			if lexer == nil {
+				lexer = lexers.Fallback
+			}
+
+			// Tokenize and format the text after stripping ANSI codes, replacing buffer if there are no errors
+			iterator, err := lexer.Tokenise(nil, ansiRegex.ReplaceAllString(buf, ""))
+			if err == nil {
+				formattedBuffer := new(bytes.Buffer)
+				if formatter.Format(formattedBuffer, style, iterator) == nil {
+					// Strip extra newline added by Chroma and replace buffer
+					buf = string(trailingNewline.ReplaceAll(formattedBuffer.Bytes(), []byte{}))
+				}
+				syntaxHighlighted = true
+			}
+		}
+
 		// Support ANSI color codes in preformatted blocks - see #59
-		if viper.GetBool("a-general.color") && viper.GetBool("a-general.ansi") {
+		// This will also execute if code highlighting was successful for this block
+		if viper.GetBool("a-general.color") && (viper.GetBool("a-general.ansi") || syntaxHighlighted) {
 			buf = cview.TranslateANSI(buf)
 			// The TranslateANSI function will reset the colors when it encounters
 			// an ANSI reset code, injecting a full reset tag: [-:-:-]
@@ -366,9 +422,21 @@ func RenderGemini(s string, width int, proxied bool) (string, []string) {
 				// Don't add the current line with backticks
 				processPre()
 
+				// Clear the language
+				lang = ""
 			} else {
 				// Not preformatted, regular text
 				processRegular()
+
+				if viper.GetBool("a-general.highlight_code") {
+					// Check for alt text indicating a language that Chroma can highlight
+					alt := strings.TrimSpace(strings.TrimPrefix(lines[i], "```"))
+					if matches := langRegex.FindStringSubmatch(alt); matches != nil {
+						if lexers.Get(matches[0]) != nil {
+							lang = matches[0]
+						}
+					}
+				}
 			}
 			buf = "" // Clear buffer for next block
 			pre = !pre