+++ sysopen/open_browser_darwin.go
@@ -7,9 +7,12 @@ import "os/exec"
 
 // Open opens `path` in default system viewer.
 func Open(path string) (string, error) {
-	err := exec.Command("open", path).Start()
+	proc := exec.Command("open", path)
+	err := proc.Start()
 	if err != nil {
 		return "", err
 	}
+	//nolint:errcheck
+	go proc.Wait() // Prevent zombies, see #219
 	return "Opened in default system viewer", nil
 }