+++ sysopen/open_browser_windows.go
@@ -8,9 +8,12 @@ import "os/exec"
 
 // Open opens `path` in default system vierwer.
 func Open(path string) (string, error) {
-	err := exec.Command("rundll32", "url.dll,FileProtocolHandler", path).Start()
+	proc := exec.Command("rundll32", "url.dll,FileProtocolHandler", path)
+	err := proc.Start()
 	if err != nil {
 		return "", err
 	}
+	//nolint:errcheck
+	go proc.Wait() // Prevent zombies, see #219
 	return "Opened in default system viewer", nil
 }