+++ webbrowser/open_browser_darwin.go
@@ -7,9 +7,12 @@ import "os/exec"
 
 // Open opens `url` in default system browser.
 func Open(url string) (string, error) {
-	err := exec.Command("open", url).Start()
+	proc := exec.Command("open", url)
+	err := proc.Start()
 	if err != nil {
 		return "", err
 	}
+	//nolint:errcheck
+	go proc.Wait() // Prevent zombies, see #219
 	return "Opened in system default web browser", nil
 }