+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Syntax highlighting for preformatted text blocks with alt text (#252, #263, [wiki page](https://github.com/makeworld-the-better-one/amfora/wiki/Source-Code-Highlighting))
+
+### Changed
+- Center text automatically, removing `left_margin` from the config (#233)
+- `max_width` defaults to 80 columns instead of 100 (#233)
+
+### Fixed
+- Modal can't be closed when opening non-gemini text URLs from the commandline (#283, #284)
+- External programs started by Amfora remain as zombie processes (#219)
+- Prevent link lines (and other types) from being wider than the `max_width` setting (#280)
+
+
 ## [1.9.2] - 2021-12-10
 ### Fixed
 - Preformatted text color showing even when `color = false` (bug since v1.8.0 at least) (#278)